feat(hooks): optimize icon hook
